Did a lot of work tonight
Got the SVG and PNG logos set up – although I may change the footer
logo.

Moved the navigation to the right hand side of the nav div.

Fixed an issue with the intro paragrah and moved it over to the left to
allow for an illustration on the homepage.

Cleaned out the content folder.

Cleaned out some of the unnecessary variables for the listing of items.

Updated the :pages partial to include all pages so that it would list
the category / shop page, and added the WePay "view cart" button.

Added the "Add to cart" button on the first product page. Prety easy
just had to grab the code from WePay and place it into the content txt
file.