Added a lot more content!
Created the two main category pages for the shop.

Created each of the custom items pages and started adding content. All
the titles are in place, now just have to add the content and preview
images

Added the Pirate Captain project page, now have to create the thumbnail
and preview images.

Added the general details text for all the prints to the _shared.txt
file so that its usable across the whole site.

Added the revised M.Davis Drawings logo for the header and the footer.

Added the homepage illustration. It may be too big to use as an svg,
still have to do some testing.

Animated the balloon on the homepage.