Getting closer to having the custom section done.
Added fonts to the project so that they load correctly in Firefox.

Fixed the width issue so that the browser doesn't think the site is
wider than it is.

Created a custom carousel css file to control how the Custom Avatar and
Custom prints section willl function.

Create a customproject template file for the custom section

Added JotForm script for testing the photo upload capabilites.