Changed the add to cart buttons on the ready made pages
Had to add some more over rides so that the buttons are more consistent
across the site and so they wouldn't move once all the css was loaded
from wepay.

Also took the price from the wepay button out and placed it in the
individual project.txt file to control its position and appearance
better.