Added keyword sorting
I knew that all I really needed was a way to find a class for each item
and then from having found that class apply a new class to hide all the
itmes that don't have that class.

Yes it requires some jquery and javascript fanciness.

I had figured out how to toggle the visibility of items but that was
the wrong behavior. Found a tutorial on Nettuts* that did exactly what
I was looking for.

----
*
http://net.tutsplus.com/tutorials/javascript-ajax/creating-a-filterable-
portfolio-with-jquery/